Detail 페이지 Response 데이터 추가
