--- src/main/java/com/idle/shoppingmall/Controller/ControllerView/OrderViewController.java
@@ -1,19 +1,0 @@
-package com.idle.shoppingmall.Controller.ControllerView;
-
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@RequiredArgsConstructor
-@Controller
-public class OrderViewController {
-
-    @GetMapping("/order")
-    String order() {
-
-
-        return "/FE/order";
-    }  // order
-
-} // end class
