+++ src/main/resources/mapper/product/product.xml
@@ -91,4 +91,32 @@
         limit #{offset}, #{size}
     </select>
 
+
+
+
+    <resultMap id="details" type="com.idle.shoppingmall.ResponseDTO.Manage.detailsObject">
+        <result column="size" property="size" />
+        <result column="amount" property="amount" />
+        <result column="sales" property="sales" />
+    </resultMap>
+    <resultMap id="ProductDetailViewDTO" type="com.idle.shoppingmall.Controller.ControllerView.ProductViews.DTO.ProductDetailViewDTO">
+        <result column="product_id" property="product_id" />
+        <result column="product_name" property="product_name" />
+        <result column="price" property="price" />
+        <result column="brand_name" property="brand_name" />
+        <result column="count_love" property="count_love" />
+        <result column="category" property="category" />
+        <result column="product_img" property="product_img" />
+        <collection property="details" ofType="com.idle.shoppingmall.ResponseDTO.Manage.detailsObject" resultMap="details"/>
+    </resultMap>
+    <select id="findViewById" resultMap="ProductDetailViewDTO">
+        SELECT p.product_id as product_id, p.pd_name as product_name, format(p.pd_price,0) as price, b.brand_name as brand_name, p.count_love as count_love,
+        p.pd_category as category, p.product_img as product_img ,upper(pd.size) as size, pd.pd_before_count as amount, pd.pd_sell_count as sales
+        from product as p
+        left outer join brand as b on p.brand_id = b.brand_id
+        left outer join product_detail as pd on p.product_id = pd.product_id
+        <where>
+            p.product_id = #{productId}
+        </where>
+    </select>
 </mapper>
