+++ src/main/resources/static/JS/DetailView/AddCart.js
@@ -5,28 +5,14 @@ CartBtn.addEventListener('click', () => {
     console.log("click")
     console.log(id)
     let count = parseInt(document.getElementById('quantity').value);
+    let size = document.getElementById('size').value
+    console.log("size = ",size)
     console.log(count)
     console.log(typeof count)
-    console.log(getSelectedSize());
     const data = {
         product_id: id,
         count: count,
-        size: getSelectedSize()
+        size: size
     }
     sendData("/api/POST/addCart", data);
 });
-
-function getSelectedSize() {
-    // 'size'라는 name 속성을 가진 모든 라디오 버튼을 가져옵니다.
-    const sizeRadios = document.getElementsByName('size');
-
-    // 가져온 라디오 버튼들을 반복하면서 체크된 항목을 찾습니다.
-    for(let i = 0; i < sizeRadios.length; i++) {
-        if(sizeRadios[i].checked) { // 만약 현재 항목이 체크되어 있다면
-            return sizeRadios[i].value; // 해당 항목의 value를 반환합니다.
-        }
-    }
-
-    // 체크된 항목이 없으면 null을 반환합니다.
-    return null;
-}
