+++ src/main/resources/templates/product_info.html
@@ -52,37 +52,28 @@
                     <img th:src="'https://kr.object.ncloudstorage.com/miniidle/storage/'+${data.product_img}" alt="Product Image" width="100%", height="100%">
                 </div>
                 <div>
-                    <h1 class="text-2xl font-bold" th:text="'제품명 : '+${data.pd_name}"></h1>
-                    <p class="text-sm text-gray-500 mt-1">SKU: 20130613</p>
-                    <p class="text-sm text-gray-500" th:text="'브랜드명 : '+${data.brand_id}"></p>
+                    <h1 class="text-2xl font-bold" th:text="'제품명 : '+${data.product_name}"></h1>
+                    <p class="text-sm text-gray-500" th:text="'브랜드명 : '+${data.brand_name}"></p>
+                    <p class="text-sm text-gray-500" th:text="'좋아요 수 : '+${data.count_love}"></p>
                     <div class="mt-4">
-                        <span class="text-3xl font-bold" th:text="'가격 : '+${data.pd_price}"></span>
+                        <span class="text-3xl font-bold" th:text="'가격 : '+${data.price}+'원'"></span>
                     </div>
                     <div class="mt-4">
-                        <h2 class="font-semibold">Color</h2>
+                        <h2 class="font-semibold" th:text="'카테고리 : '+${data.category}">Color</h2>
                         <div class="flex space-x-2 mt-2">
                             <!-- Color options here -->
                         </div>
                     </div>
                     <div class="mt-4">
-                        <h2 class="font-semibold">Size</h2>
-                        <div class="flex space-x-4 mt-2">
-                            <button class="size-btn hover:bg-pink-100 rounded-full w-8 h-8 flex items-center justify-center" data-size="XS">
-                                <span class="text-gray-700 text-lg">XS</span>
-                            </button>
-                            <button class="size-btn hover:bg-pink-100 rounded-full w-8 h-8 flex items-center justify-center" data-size="S">
-                                <span class="text-gray-700 text-lg">S</span>
-                            </button>
-                            <button class="size-btn hover:bg-pink-100 rounded-full w-8 h-8 flex items-center justify-center" data-size="M">
-                                <span class="text-gray-700 text-lg">M</span>
-                            </button>
-                            <button class="size-btn hover:bg-pink-100 rounded-full w-8 h-8 flex items-center justify-center" data-size="L">
-                                <span class="text-gray-700 text-lg">L</span>
-                            </button>
+                        <h2 class="font-semibold">사이즈</h2>
+                        <div class="flex space-x-2 mt-2">
+                            <select  name="size" id="size" class="block w-full bg-gray-100 border border-gray-300 rounded py-2 px-4 focus:outline-none focus:bg-white focus:border-purple-500">
+                                <option th:each="detail : ${data.details}" th:value="${detail.size}" th:text="'사이즈 : '+ ${detail.size}+'      남은 수량 : '+${detail.amount}"></option>
+                            </select>
                         </div>
                     </div>
                     <div class="mt-4">
-                        <h2 class="font-semibold">Quantity</h2>
+                        <h2 class="font-semibold">갯수</h2>
                         <div class="flex space-x-2 mt-2">
                             <button id="decrement" class="bg-gray-200 hover:bg-gray-300 rounded-full w-8 h-8 flex items-center justify-center">
                                 <span class="text-gray-700 text-lg">-</span>
@@ -211,29 +202,6 @@
 <script src="/JS/DetailView/printComment.js"></script>
 <script src="/JS/DetailView/addComment.js"></script>
 <script>
-    // 사이즈 선택 버튼 클릭 이벤트 처리
-    const sizeButtons = document.querySelectorAll('.size-btn');
-
-    sizeButtons.forEach(button => {
-        button.addEventListener('click', () => {
-            // 클릭한 버튼이 이미 핑크색 클래스를 갖고 있는지 확인
-            const isPink = button.classList.contains('bg-lightpink-600');
-
-            // 모든 버튼의 색상을 회색으로 변경
-            sizeButtons.forEach(btn => {
-                btn.classList.remove('bg-lightpink-100');
-                btn.classList.add('bg-gray-200');
-            });
-
-            // 클릭한 버튼이 핑크색 클래스를 갖고 있지 않으면 핑크색 클래스 추가
-            if (!isPink) {
-                button.classList.remove('bg-gray-200');
-                button.classList.add('bg-lightpink-600');
-            }
-        });
-    });
-
-
     // 수량 업데이트 함수
     function updateQuantity(change) {
         const quantityInput = document.getElementById('quantity');
